vasco-dasch/pyproject.toml
[project]
name = "vasco-dasch"
version = "0.1.0"
description = "VASCO x DASCH cross-match pipeline"
authors = [
    {name = "Jamie"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "daschlab>=1.0.0",
    "astropy>=6.0",
    "photutils>=1.13",
    "requests>=2.32",
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.13",
    "matplotlib>=3.9",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "openpyxl (>=3.1.5,<4.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
